fix with copy url and preview for folders and files

+++ src/components/google-drive/ContextMenu.jsx
@@ -1,13 +1,12 @@
 import React, { useState } from "react";
 import { Button, Form, Modal } from "react-bootstrap";
 import { Link } from "react-router-dom";
+import { useAuth } from "../../context/AuthContext";
 import { db } from "../../firebase";
 import styles from "./ContextMenu.module.css";
 
 export default function ContextMenu({
   show,
-  x,
-  y,
   url,
   path,
   urlRef,
@@ -19,65 +18,44 @@ export default function ContextMenu({
   function copyToClipboard(path) {
     navigator.clipboard.writeText(path);
   }
+  const { currentUser } = useAuth();
 
   const toggleModal = () => setShowModal((prevState) => !prevState);
+  const folderUrl = window.location.origin + url;
 
-  const submitHandler = e => {
+
+  // console.log('url', url)
+  // console.log('folderUrl', folderUrl)
+
+  const submitHandler = (e) => {
     // prevent form submitting form
     e.preventDefault();
 
     // send requset for deleting
-    db.deleteFile(path, id);
-  }
+    if (file) db.deleteFile(path, id);
+    else db.deleteFolder(path, currentUser.uid);
+  };
 
   return (
     <>
       <div
         className={`${styles.wrapper} ${show ? styles.visible : styles.hidden}`}
-        style={{
-          left: `${x}px`,
-          top: `${y}px`,
-        }}
         {...props}
       >
         <div className={`${styles.content}`}>
           <ul className={`${styles.menu}`}>
             <li className={`${styles.item}`}>
-              {file && (
-                <a
-                  className="w-100"
-                  target="_blank"
-                  href={url}
-                  rel="noreferrer"
-                >
-                  <i className={`uil uil-eye`}></i>
-                  <span>Preview</span>
-                </a>
-              )}
-              {!file && (
-                <Link
-                  className="w-100"
-                  target={"_blank"}
-                  to={url}
-                  onClick={(e) => {
-                    console.log("clicked");
-                  }}
-                >
-                  <i className={`uil uil-eye`}></i>
-                  <span>Preview</span>
-                </Link>
-              )}
+              <a className="w-100" target="_blank" href={url} rel="noreferrer">
+                <i className={`uil uil-eye`}></i>
+                <span>Preview</span>
+              </a>
             </li>
             <li
               className={`${styles.item}`}
-              onClick={() => copyToClipboard(url)}
+              onClick={() => copyToClipboard(file ? url : folderUrl)}
             >
               <i className={`uil uil-link-alt`}></i>
               <span>Get Link</span>
-            </li>
-            <li className={`${styles.item}`}>
-              <i className={`uil uil-edit`}></i>
-              <span>Rename</span>
             </li>
             <li className={`${styles.item} text-danger`} onClick={toggleModal}>
               <i className={`uil uil-trash-alt`}></i>
@@ -90,10 +68,14 @@ export default function ContextMenu({
         <Modal.Header>Delete File</Modal.Header>
         <Form onSubmit={submitHandler}>
           <Modal.Body className="text-center">
-            Are you sure you want to delete this file?
+            Are you sure you want to delete this {file ? "file" : "folder"}?
           </Modal.Body>
           <Modal.Footer>
-            <Button type="submit" variant="outline-danger" onClick={toggleModal}>
+            <Button
+              type="submit"
+              variant="outline-danger"
+              onClick={toggleModal}
+            >
               Delete
             </Button>
             <Button variant="secondary" onClick={toggleModal}>
